Adding more changes

+++ tests/testHelpers.js
@@ -1,18 +1,49 @@
 const mongoose = require('mongoose');
-const { MongoMemoryServer } = require('mongodb-memory-server');
-
 let mongoServer;
+let usingExternalMongo = false;
+let patchedListen = false;
+
+const forceLocalhostListen = () => {
+  if (patchedListen) return;
+  const net = require('net');
+  const originalListen = net.Server.prototype.listen;
+  net.Server.prototype.listen = function (...args) {
+    if (typeof args[0] === 'number') {
+      const hasHost = typeof args[1] === 'string';
+      if (!hasHost) {
+        return originalListen.call(this, args[0], '127.0.0.1', ...args.slice(1));
+      }
+    }
+    return originalListen.apply(this, args);
+  };
+  patchedListen = true;
+};
 
 const connect = async () => {
-  mongoServer = await MongoMemoryServer.create();
+  const externalUri = process.env.MONGODB_URI;
+  if (externalUri) {
+    usingExternalMongo = true;
+    await mongoose.connect(externalUri);
+    return;
+  }
+
+  forceLocalhostListen();
+  const { MongoMemoryServer } = require('mongodb-memory-server');
+  mongoServer = await MongoMemoryServer.create({
+    instance: { ip: '127.0.0.1' },
+  });
   const uri = mongoServer.getUri();
   await mongoose.connect(uri);
 };
 
 const closeDatabase = async () => {
-  await mongoose.connection.dropDatabase();
+  if (!usingExternalMongo) {
+    await mongoose.connection.dropDatabase();
+  }
   await mongoose.connection.close();
-  await mongoServer.stop();
+  if (mongoServer) {
+    await mongoServer.stop();
+  }
 };
 
 const clearDatabase = async () => {
